pyproject.toml
[project]
name = "llm-agent"
version = "0.1.0"
description = "Explore about LLM Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "langchain-google-genai>=2.0.7",
    "langchain>=0.3.13",
    "langgraph>=0.2.60",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "ragas>=0.2.9",
    "scipy>=1.14.1",
    "streamlit>=1.41.1",
]
